Implement the "query" and "fragment" response modes.
When "query" mode is specified, the "code" and "state" parameters of the
redirect_uri URI are replaced with the grant token and state parameter.

When "fragment" mode is specified, redirect_uri's fragment identifier is
replaced with the form URL-encoded grant token and state parameter.

This commit also moves the templates into a separate "oauth" folder
since they are not part of the accounts forms.