Add POST /oauth/token route for trusted OAuth clients
Add a route that trusted OAuth clients can use to get an access and
refresh token if the user is already logged in to the Hypothesis web
service. This will enable trusted clients to seamlessly & automatically
log in with the same user account they used to sign in to the web
service.

The client is verified by checking the Origin header in the request
against the pre-registered Origin header for that client. The use of the
POST verb is required to ensure that even clients which are same-origin
with the service send this header in the request.