Color border of input fields with errors correctly in lists
Set the color of the input field border based only on classes applied to
it directly, rather than class of the parent `.form-input`, which is
part of `mapping_item` and thus not present in list field items rendered
by `sequence_item`.

Update the logic in `FormInputController` to remove the class applied
directly to the `<input>` when the user edits the input, as well as the
class applied to the containing "mapping_item".