Add organization into group service
Add organization as an optional input when creating each of the
different groups. Organization is defaulted to the `__default__`
organization. Note this uses an organization_fetcher that goes out
and grabs the organization based on it's public id. This is nice
both for specifying the default org directly in the param def,
in the cli when the user needs to pass in the org via the command
line (aka they can't create a python object), and in the admin page
where it would be simpler to grab the pubid from a list of orgs
and pass it from the front end to the service.