Allow None in organization validate logo method
The logo for an organization can be None thus the validate logo
method needs to be able to accept that as a valid option. Before,
it was assumed that the default was None so None would never be
run through the validate logo method. With the introduction of
the create organization service this may no longer be the case.